tedla-hypertension-project-databricks/requirements.txt
annotated-types==0.7.0
attrs==25.4.0
blis==0.7.11
# CacheControl==0.12.11  # Poetry dependency, not needed locally
catalogue==2.0.10
certifi==2025.11.12
charset-normalizer==3.4.4
# cleo==2.0.1  # Poetry dependency
click==8.3.1
cloudpathlib==0.23.0
colorama==0.4.6
confection==0.1.5
# crashtest==0.4.1  # Poetry dependency
cymem==2.0.13
Cython==3.1.7
# dulwich==0.20.50  # Poetry/Git dependency
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
html5lib==1.1
idna==3.11
importlib-metadata==6.0.0
iniconfig==2.3.0
jaraco.classes==3.2.3
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
# keyring==23.13.1  # Poetry dependency
langcodes==3.5.0
language_data==1.3.0
loguru==0.7.3
marisa-trie==1.3.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
medspacy==1.3.1
medspacy_quickumls==3.2
medspacy_unqlite==0.9.8
more-itertools==9.0.0
murmurhash==1.0.15
# nlp_preprocessor==0.0.1  # Databricks-only, not needed locally
nltk==3.9.2
numpy==1.26.4
packaging==25.0
pandas>=2.0
# pexpect==4.8.0  # Unix-only, not needed on Windows
# pkginfo==1.9.6  # Poetry dependency
platformdirs==2.6.2
pluggy==1.6.0
polars>=0.20
# poetry==1.3.2  # Build tool, not needed at runtime
# poetry-core==1.4.0
# poetry-plugin-export==1.2.0
preshed==3.0.12
# py4j>=0.10.9  # Databricks-only, not needed locally
pyarrow==22.0.0
pydantic==2.12.4
pydantic_core==2.41.5
PyFastNER==1.0.10
Pygments==2.19.2
pyrsistent==0.19.3
PyRuSH==1.0.12
pysbd==0.3.4
pysimstring==1.3.0
# pyspark>=3.4  # Databricks-only, not needed locally
# pytest==9.0.1  # Dev dependency, not needed at runtime
python-dateutil>=2.8
python-dotenv==1.2.1
pytz>=2023.3
quicksectx==0.4.1
rapidfuzz==2.13.7
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-toolbelt==0.10.1
rich==14.2.0
rpds-py==0.29.0
# shellingham==1.5.0.post1  # Poetry dependency
six>=1.16
smart_open==7.5.0
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.2
thinc==8.2.5
tomlkit==0.11.6
tqdm==4.67.1
typer==0.20.0
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata>=2023.3
Unidecode==1.4.0
urllib3==2.5.0
# virtualenv==20.17.1  # Not needed at runtime
wasabi==1.1.3
weasel==0.4.3
win32_setctime==1.2.0
wrapt==2.0.1
zipp==3.11.0
